Allow switch port timeout to be configured
Previously we would timeout switch ports after 300s of inactivity.
Ideally active connections would be configured to send keep-alive packets
sooner than this. To better support cases with no keep-alives, it's
useful to be able to reconfigure the idle timeout (e.g. to make it
much bigger).

This patch adds

- a command-line argument `--port-max-idle-time`
- a database key `slirp/port-max-idle-time`

which allow the default 300s to be overriden.

Related to #235
Related to #234

Signed-off-by: David Scott <dave.scott@docker.com>