Add support for disconnecting with a reason
Adds a new Response packet that may be sent as a response to the Command
sent by the client.  The Response currently either contains the virtual
interface description or a disconnect message and a string with a
reason. This allows us to return more error messages in the future, such
as when an IP has already been used by a different UUID or we run out of
IPs.

The new packet is currently only sent following an Ethernet or
Preferred_ipv4 command from the client. If an error occurs earlier in
the negotitation (e.g. during version check), this is not reported back
to the client with a Response message. Instead, the client should be
able to determine that the negotiation failed, as the server will
respond with its own version (which will not match the client's) before
disconnecting.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>