Enable TCP keepalives in the internal link
In cases where the TCP connection state in the client disappears (maybe
because a container is torn down and the IP address and NAT tracking
state is deleted at the same time) we wish to be able to close the
external connection.

The internal link is very low latency and therefore a short keepalive
timeout is possible.

Signed-off-by: David Scott <dave@recoil.org>