Datagram.Udp.read: emulate a connected socket
Previously we would call `recvfrom` and accept any traffic even if
this socket was supposed to be connected to a specific remote address.

This patch drops packets in `read` which have missing or invalid
from addresses. Note this code is not used by the UDP NAT
implementation: the NAT uses `recvfrom` to deliberately receive
traffic from other addresses.

Signed-off-by: David Scott <dave.scott@docker.com>