Create and use a Configuration.t record
Previously we would use an ad-hoc mix of global variables and database
watchers. This code formalises the configuration into a record which
is populated from the database but which could be populated from
somewhere else (e.g. a configuration file) later.

This patch also removes the automatic restart on configuration change--
we would rather a higher-level tool would perform a graceful shutdown.

Signed-off-by: David Scott <dave.scott@docker.com>