icmp: receive incoming TTL exceeded and forward
An ICMP datagram with TTL exceeded contains a truncated copy of
the original IPv4 datagram. In our case this will contain a truncated
ICMP echo request. This patch parses the packet and rewrites both
the destination and the id in the nested copy because

- the destination we receive from the host will have the host's IP
  in it, we need to use the VM's IP
- the id we receive from the host is chosen by us in the NAT
  translation, so we have to translate it back

With these changes, ICMP traceroute works.

Signed-off-by: David Scott <dave.scott@docker.com>