build: allow the system OCaml compiler to be used
Previously we always supplied a `--comp` argument to `opam init` and
would therefore build a custom OCaml compiler -- this is fine for the
CI case where we wish to control the version precisely.

This patch interprets an empty `OPAM_COMP` environment variable as
meaning "use the system compiler" which is useful for (eg) the homebrew
case.

Signed-off-by: David Scott <dave.scott@docker.com>