Add an argument --dhcp-path <filename>
A DHCP response has options which should include

- search domains
- the domain name

When we're using the "Host" resolver there is no convenient place to
put these configuration variables. This patch adds a `--dhcp-path <filename>`
argument which watches the path `<filename>`, parses it and updates
the DHCP server configuration. Note the client will have to renew the
lease before it observes the difference.

The file should look like this:

{"searchDomains":[],"domainName":"cam.docker.com"}

Signed-off-by: David Scott <dave.scott@docker.com>