Don't block DNS queries in Uwt_preemptive
DNS requests can block for a long time, particularly if the upstream
server is broken. The preemptive pool is quite small by default
(4 threads) so it's very easy to fill it up with slow-to-fail DNS queries.

This patch switches to a more async style using `Uwt.Poll` to query the
result file descriptor for readability, and only uses `Uwt_preemptive`
to read the final result.

Signed-off-by: David Scott <dave.scott@docker.com>