Remove the "extra_dns" IPs
These were a remnant of an earlier attempt to make DNS more reliable.
The idea was to map the "extra_dns" IPs onto the real external DNS IPs
and use the VM's resolvers to query them. Unfortunately this failed
because many common resolvers (including the musl one) only use the
first 3 IPs and a Mac/Windows host may have more than that.

On Windows we use a resolver library and on Mac we use the macOS APIs
directly, exposing both as a single DNS IP. This code is redundant and
can be removed.

Signed-off-by: David Scott <dave.scott@docker.com>