HTTP proxy: fix connection handling
Previously we would keep reading requests from a connection until we hit
a hard error. This patch modifies the `fetch_direct` function to return
a `bool` which is `true` if we should try to keep going, and `false` if
we should close the connection immediately.

If we encounter a parse error, then we should close the connection.

Once we have finished proxying binary data in HTTP CONNECT, we should
close the connection.

In the case where we have successfully completed a regular HTTP
transaction (GET, PUT, POST etc) then we try optimistically to keep the
connection open and proxy more requests.

Signed-off-by: David Scott <dave.scott@docker.com>