build: `make` by default will use the custom package repository
Previously there was a difference between `make` run from the CI and
`make` run by a developer: the CI would use the carefully managed
package universe while the developer build would be against whatever
library versions happen to be installed.

This patch changes the developer default to be the same as the CI.
The targets depend on `$(REPO_ROOT)/_build/opam` and run the setup script
`./scripts/depend.sh` to create it. All build commands are prefixed with
`opam config exec` to ensure they use the correct library versions.

After this change to deliberately compile against a different package
universe, simply set the `OPAMROOT` environment variable. This means
that experienced `opam` users will have to `unset OPAMROOT` in order
to make a standard build, but it means that inexperienced `opam` users
find that `make` should just work by default.

Signed-off-by: David Scott <dave.scott@docker.com>