Remove the TCP-only codepath
Previously we used a TCP-level proxy when the system had a proxy configured.
Unfortunately this doesn't work when a proxy requires authorization -- we
need to parse the requests and add/replace the `Proxy-Authorization` header.

This patch removes the TCP-level proxy and processes all traffic through the
regular proxy -- the one used before when there was no upstream proxy defined.

This patch adds a notion of a `route` which describes
- the address of the next hop
- whether the next hop is an origin server or a proxy
- the host (of the server with the resource)
- the port (of the server with the resource)
- a description string

Requests are parsed, a `route` is computed and then the `fetch` function
either queries the origin server or the upstream proxy depending on
- whether a proxy is configured
- the request URI, if present
- the host: header, if present
- the proxy excludes

This patch includes the following minor changes:

- Extract a common function `proxy_bytes` which proxies bytes between
  two buffered channels.
- The common `proxy_request` now knows how to perform `HTTP CONNECT too.
- The `proxy_request` now returns a boolean saying whether the connection
  should be kept open or closed.
- Extract `add_proxy_authorization` helper function to manage the
  `Proxy-Authorization` header.
- Extract `send_error` function for replying with HTTP errors.
- Add a helper function `address_of_proxy`.

Signed-off-by: David Scott <dave.scott@docker.com>