dns: return NXDomain for a builtin where we don't have RRs
Previously the client could request a builtin name like
`docker.for.mac.http.internal` and ask for both A and AAAA
records. We would return the A record and then potentially
timeout the AAAA response if the upstream DNS server is missing
or unresponsive. This manifests, for example, as very slow HTTP
transactions.

This patch claims all RRs for the "builtin" names. If we don't
actually have an RR (e.g. no IPv6 address) then we return
NXDomain and don't attempt to forward the request upstream.

Signed-off-by: David Scott <dave.scott@docker.com>