HTTP proxy: handle localhost
HTTP traffic on port 80 or 443 are dispatched based on destination IP
address:
- if to "localhost" they are forwarded to 127.0.0.1
- if to anywhere else they are funnelled through the transparent proxy

HTTP traffic sent to vpnkit's gateway IP on port 3128 (the HTTP proxy)
but whose final destination is one of the names for the host's localhost
is unfortunately routed to the wrong place.

This patch adds a `localhost_names` argument to the HTTP proxy `route`
function. If the destination host is one of the `localhost_names` then
it is rewritten as `localhost` which will then cause connections to the
host's `localhost`.

Signed-off-by: David Scott <dave.scott@docker.com>