HTTP: consult the "localhost" names and IPs in the transparent proxy
Previously we would special-case names like

  host.internal

in the regular proxy path but we missed the transparent proxy path.
If the proxy is set to

  http://host.internal:8080/

then we would intercept outgoing traffic to port 80 and perform a DNS lookup
on `host.internal` which would fail.

This patch makes sure we consult the list of "localhost" names in the
transparent path.

Signed-off-by: David Scott <dave.scott@docker.com>