test: check the proxy respects Connection: close from origin servers
This test starts an origin server on localhost which will respond with
a minimal HTTP/1.0 response, with no connection-length or transfer-encoding.
A client connects to the builtin HTTP proxy on port 3128 and requests
a URI on the origin server. The test requires the body to be proxied
correctly by the proxy (i.e. all data up to EOF should be transferred)

Tests #361

Signed-off-by: David Scott <dave.scott@docker.com>