Add a notion of "UDPv4 forwards" for the gateway IP
`udpv4_forwards` is a table mapping a source port to a destination
IP address and port. Traffic sent to the gateway will be matched
against this list, and NAT'ed.

This patch also generalises the `Udp_nat` module to allow the source
port of replies to be set to either the outside source port (the default)
or the port the client originally sent the UDP packet to. The default is
more general but when "forwarding" we wish to act more like a proxy and
reply from the original destination port.

Note also the `Gateway` module must construct a `Udp_nat` instance and
`send_reply` function because the traffic must come from the gateway
address and not the "remote" address.

Signed-off-by: David Scott <dave.scott@docker.com>