Reduce ICMP log spam
Unfortunately the ICMP sockets receive more traffic than we expect, for
example if someone opens a terminal on the host and types `ping 8.8.8.8`
then `vpnkit` will receive this traffic and complain that it didn't expect
it.

This patch sets the "unexpected" log messages to "debug" level so they will
be dropped by default.

Signed-off-by: David Scott <dave.scott@docker.com>