ICMP: don't log an error if we can't parse an ICMP packet
The ICMP sockets on the host are unusual in that they receive all the
traffic seen by the host. For example if the user opens another terminal
and runs `ping`, then `vpnkit` will see the responses too even though
it didn't send the requests.

It's safe to ignore ICMP packets that we don't recognise since they
weren't meant for us. This patch reduces the log level from `error`
to `debug`.

Fixes #381

Signed-off-by: David Scott <dave.scott@docker.com>