vpnkit-tap-vsockd: add a --post-up-script <script>
Previously the image was `FROM scratch` and the network interface
was created asynchronously, making it hard to configure the IP address.

This patch makes the image `FROM alpine:3.6` so that a basic scripting
environment is available and adds the argument

```
  --post-up-script <script>
```

where <script> will be run as soon as the ethernet device exists.

Signed-off-by: David Scott <dave.scott@docker.com>