Add `--port-forwards` as a synonym for `--vsock-path`
This parameter is currently used on the Mac for naming the path to
the virtio-vsock connection socket, used when forwarding ports into
the VM.

In a future patch we need to extend it to work also with Hyper-V
sockets.

This patch adds a synonym, keeping the old parameter for backwards
compatibility. The variable `vsock_path` is renamed to `port_forwards`
inside the code.

Signed-off-by: David Scott <dave.scott@docker.com>