Use --vsock-path / --port-forwards on both Windows and Mac
Previously on Windows we used the 9P port filesystem server URL for
both the 9P server control plane and the data plane (connected to
vpnkit-forwarder).

Unfortunately now we switch to VM -> host connection for the 9P server,
the command-line doesn't contain the VM's GUID so port forwarding is
broken.

This patch uses the existing --vsock-path / --port-forwards argument
to configure the data plane on both Windows and Mac. For backwards
compatibility it still supports a raw path, but now also supports

  file:///foo/bar
  hyperv-connect://vmid/serviceid/

Signed-off-by: David Scott <dave.scott@docker.com>