Remove the Hyper-V socket "shutdown" protocol
Previously we used a custom protocol on top of Hyper-V sockets because
they didn't support `shutdown`. With modern Windows and modern Linux the
`shutdown` works (at least when the connections are initiated from the
host) and therefore the Go `linuxkit/virtsock` library has dropped the
custom protocol. Since we're talking to `vpnkit-forwarder` using
`linuxkit/virtsock` we must drop the custom protocol too.

This uses an experimental version of `mirage/ocaml-hvsock` with support
for syscall-based `shutdown`.

Signed-off-by: David Scott <dave.scott@docker.com>