vmnet: prevent the main listener from being cancelled after 4s
The DHCP client has a 4s timer[1] and when it expires it calls
`Lwt.cancel` on the `Net.listen` thread[2]. In the end-to-end
tests the cascade of thread cancellations causes the `read` call
to return with `Lwt.Cancelled`, the file descriptor is closed
and everything falls apart.

This patch catches the `Lwt.Cancel` and logs a warning rather than
tearing down the connection.

[1] https://github.com/mirage/charrua-core/blob/d6d8736e1c5a42782c4884cea4f86059acfde3ca/client/lwt/dhcp_client_lwt.ml#L16
[2] https://github.com/mirage/charrua-core/blob/d6d8736e1c5a42782c4884cea4f86059acfde3ca/client/lwt/dhcp_client_lwt.ml#L85

Signed-off-by: David Scott <dave.scott@docker.com>