Add support for more Hypervisor socket URIs
Previously we had hard-coded support for `hvsock-listen`, `hvsock-connect`
and `fd:4`. This patch adds support for other socket URIs by falling
back to the `Hvsock.Socket` "generic" API. In particular this supports
listening on a AF_VSOCK on the Mac, which will allow us to make the
behaviour on Linux and Mac more similar.

Signed-off-by: David Scott <dave.scott@docker.com>