go: multiplexer: support net.Conn
Previously we supported io.Reader, io.Writer, io.Closer, ReadClose() and
WriteClose(). However it's useful to conform to the whole net.Conn interface
since some libraries require it, notably the http client, the docker client
and net.Dialer.

This bulk of this patch adds read and write deadlines. The local and remote
address functionality is stubbed as it's not particularly meaningful.

Signed-off-by: David Scott <dave.scott@docker.com>