pkg/libproxy: avoid a Write/CloseWrite or Write/Close race
Previously we would send the `Close` or `Shutdown` message before
setting the `closeSent` or `shutdownSent` variables which means
that a `Write` call could race with us, check the variables,
conclude the connection is still open and write a payload, which
would arrive after the remote has forgotten about the connection.

In this patch we set the `closeSent` or `shutdownSent` variables
to true before sending the `Close` or `Shutdown`. A `Write`
call which arrives (or unblocks) after this point is guaranteed
to return EOF rather than sending data.

Signed-off-by: David Scott <dave.scott@docker.com>