Add a notion of "TCPv4 forwards" for the gateway IP
`tcpv4_forwards` is a table mapping a source port to a destination
IP address and port. Traffic sent to the gateway will be matched
against this list, and proxied at the TCP level.

This is the TCP analog of 66a4628ccc3f5dd970fa79e3e5b512e3b54385d5

Related to #375

Signed-off-by: David Scott <dave.scott@docker.com>