UDP: copy packets before forwarding
The lower part of the stack does maintain references to the packets
(e.g. the .pcap capturing layer) so it's important not to have each
packet share the same buffer.

This patch also makes it clearer that `loop` is processing packets
from the outside going into the virtual network, and the receive
buffer is private.

Signed-off-by: David Scott <dave.scott@docker.com>