go mux: fix double-free connection ID race
The connection ID is free when
- we have called `Close` and sent the `Close` message
- we have received the `Close` message from the remote

Previously it was possible for both these conditions to become true
at once in different goroutines, which would result in the free
happening twice. This probably didn't result in an observable error
since it's very unlikely that the very same connection ID was
immediately re-used, since we have a counter which sweeps its way
through the entire 2^32 connection ID space.

This patch associates a reference count with the connection which is
initially 2: 1 for the local and 1 for the remote. When we send or
receive a `Close` we decrement the reference count and check it *with
the multiplexer mutex held*.

Signed-off-by: David Scott <dave.scott@docker.com>