ocaml mux: fix double close
Previously it was possible to break the system by calling `Close`
twice. This patch uses the `close_sent` flag to prevent this and
switches to a reference count to decide when to free the connection
id. Note it's not possible to use the same boolean for both purposes
as we need to set it early to prevent double send of `Close` but we
need to set it late to prevent double free of the id.

Signed-off-by: David Scott <dave.scott@docker.com>