go: don't return transient Unix domain socket forward errors
We assume Unix domain sockets are for forwarding to internal
services which take a few seconds to start. Rather than fail and
require the client to retry, we will catch the "connection
refused" and retry every 5s for 120s-- this should be enough
to hide the gap while the VM is booting.

Signed-off-by: David Scott <dave.scott@docker.com>