Fix port Unexpose of INADDR_ANY ports
If a client requests any port (== 0 i.e. INADDR_ANY) the port is resolved
to a concrete port when it is bound.

Previously an attempt to list this port and unexpose it would leak because
we wouldn't find the concrete port in the map.

This patch ensures that the concrete port is stored so the port can be
unexposed properly. This also fixes a potential problem where two forwards
of any port would only result in one actual forward, because the second one
would be assumed to be a duplicate.

Signed-off-by: David Scott <dave@recoil.org>