Set O_CLOEXEC on received file descriptors
By default Go sets O_CLOEXEC on file descriptors it creates. This
ensures that they aren't accidentally inherited over fork+exec.

Unfortunately when we `recvmsg` a file descriptor, there is no
flag on Darwin to set the close on exec flag atomically. We must
do it ourselves and tolerate the small race window where the fd
could be inherited.

Signed-off-by: David Scott <dave@recoil.org>