vpnkit-forwarder: avoid spinning when connecting data channel
There is currently no initial handshake to check whether the remote
is actually there, so we can connect() successfully but then immediately
read an EOF in some cases (e.g. port forwarding when the server isn't
running). Previously we would spin in a tight loop. This patch adds
a log and a 1s delay to avoid spinning.

Signed-off-by: David Scott <dave@recoil.org>