libproxy: don't fail if Close messages arrive out-of-order
A test run demonstrated a two-way Close
```
[2023-04-04T05:48:03.965050200Z][com.docker.backend.exe][I] recv  3 Close
[2023-04-04T05:48:03.965050200Z][com.docker.backend.exe][I] send  3 Close
```
at this point the channel refcount is 0 and the channel considered closed
```
[2023-04-04T05:48:03.965050200Z][com.docker.backend.exe][I] close 3 -> UDP:127.0.0.1:5201
```
however an in-progress Read call read some buffered data and sent a window advertisement:
```
[2023-04-04T05:48:03.965050200Z][com.docker.backend.exe][I] recv  3 Window 935323206
```
which broke the multiplexer with
```
[2023-04-04T05:48:03.918050500Z][com.docker.backend.exe][I] Multiplexer main loop failed with Unknown channel id 3 Window 935323206
```

Previously it was illegal to send anything after a Close(). However it's legal
for clients to call Close() in goroutines while other goroutines are calling
Read() or Write(). Each send call competes for the writeMutex, which means it's
possible for the messages to be sent out-of-order.

Since we have 2**32 channel numbers there should not be any confusion if we
simply expect the occasional out-of-order packet after a Close and drop it.

Signed-off-by: David Scott <dave@recoil.org>